F-115: androidUtilTarget Path B helper + NiA Room dogfood docs
Add androidUtilTarget (resourcesTarget twin) so derived util kinds can
own plugins. Document NiA spike Room Path B (roomAndroidUtil) green.
