+++ plugins/android/src/main/java/androidUtil.kt
@@ -1,22 +1,10 @@
-import tools.forma.android.feature.AndroidLibraryFeatureConfiguration
-import tools.forma.android.feature.androidLibraryFeatureDefinition
-import tools.forma.android.feature.applyFeatures
-import tools.forma.android.feature.kotlinAndroidFeatureDefinition
-import tools.forma.android.target.AndroidTargetRegistry
 import tools.forma.android.target.AndroidTargetTypes
 import tools.forma.owners.NoOwner
 import tools.forma.owners.Owner
-import tools.forma.android.validation.disallowResources
 import tools.forma.android.visibility.Public
 import tools.forma.android.visibility.Visibility
 import org.gradle.api.Project
-import tools.forma.android.feature.processorConfigurationFeatures
-import tools.forma.deps.core.applyDependencies
-import tools.forma.deps.core.applyTargetPlugins
 import tools.forma.deps.core.FormaDependency
-import tools.forma.deps.fleet.registerFormaLayout
-import tools.forma.validation.asValidator
-import tools.forma.validation.validate
 
 /**
  * TODO
@@ -33,6 +21,7 @@ import tools.forma.validation.validate
  * )
  * ```
  *
+ * Path B derived util kinds (Room, etc.) use [androidUtilTarget] with a derived type.
  */
 fun Project.androidUtil(
     packageName: String,
@@ -43,29 +32,13 @@ fun Project.androidUtil(
     /** Enable Jetpack Compose; defaults to project-wide `compose` setting. */
     compose: Boolean = Forma.settings.compose,
 ) {
-
-    disallowResources()
-
-    //TODO unify with util, use androidJar dependency
-    AndroidTargetRegistry.selfValidator(AndroidTargetTypes.androidUtil).asValidator().validate(target)
-    registerFormaLayout(packageName)
-
-    val androidFeatureConfig = AndroidLibraryFeatureConfiguration(
+    androidUtilTarget(
+        type = AndroidTargetTypes.androidUtil,
         packageName = packageName,
-        compose = compose,
-    )
-
-    applyFeatures(
-        androidLibraryFeatureDefinition(androidFeatureConfig),
-        kotlinAndroidFeatureDefinition()
-    )
-    val processors = processorConfigurationFeatures()
-    applyTargetPlugins(AndroidTargetTypes.androidUtil, configurationFeatures = processors)
-
-    applyDependencies(
-        validator = AndroidTargetRegistry.validatorFor(AndroidTargetTypes.androidUtil).asValidator(),
+        owner = owner,
+        visibility = visibility,
         dependencies = dependencies,
         testDependencies = testDependencies,
-        configurationFeatures = processors
+        compose = compose,
     )
 }
